Remove type
